requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
jinja2==3.1.2
python-binance==1.0.19
matplotlib==3.8.2
pydantic==2.5.0
langchain
langchain-openai
langchain-community
openai>=1.10.0 
